Fix server remote start
Increase the actor timeout to 10s. This is actually a stupid hack. We know the
each message will probably not take longer than 2s to process, so our timeout
is basically 2s*(number of servers). But this isn't scalable. We probably
want to use a router actor here for better paralellism.

Fix start button event handling. We weren't properly binding the event, so it
wasn't picking up clicks on newly added button elements.

Fix a typo in the sever info selector.

Reduce the timeout for querying the server to 10s after start. 2m is just way
too long. 10s probably isn't enough for a modded server, So this may be worth
revising later.