Refactor Server config object and start method
Move the ServerConfig case class into its own file, renaming it to Server.
Add an additional remoteUser property to the Server case class, with a helper
method to determine if a server can be started, and to start the server.

Also factor out the scala-ssh library in favor of jsch. This lets us provide
a private key via variable, which means we can load it from an environment var,
so we can deploy to Heroku.